Animating logic in the process

+++ app/src/main/java/com/example/jeu2048/gameRender/CellRenderer.java
@@ -3,5 +3,5 @@ package com.example.jeu2048.gameRender;
 import android.graphics.Canvas;
 
 public abstract class CellRenderer {
-    public abstract void drawCell(Canvas canvas, int top, int left, int cellWidth, int cellHeight, long value);
+    public abstract void drawCell(Canvas canvas, float top, float left, float cellWidth, float cellHeight, long value);
 }
